style: clean up
